confirmed all the endpoints are working

+++ mangav2/src/main/java/com/example/mangav2/service/UserService.java
@@ -14,9 +14,9 @@ public interface UserService extends UserDetailsService {
 
     public User getUser(String username);
 
-    public Iterable<Mangas> addMangasToUserList(String username, int manga_id);
+    public Iterable<Mangas> addMangasToUserList(String username, Long manga_id);
 
-    public Iterable<Mangas> deleteMangasFromUserList(String username, int manga_id);
+    public Iterable<Mangas> deleteMangasFromUserList(String username, Long manga_id);
 
     public List<Mangas> listUserMangaList();
 }
